+++ x/tokenfactory/client/cli/tx_mint.go
@@ -1,0 +1,52 @@
+package cli
+
+import (
+	"strconv"
+
+	"github.com/cosmos/cosmos-sdk/client"
+	"github.com/cosmos/cosmos-sdk/client/flags"
+	"github.com/cosmos/cosmos-sdk/client/tx"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/spf13/cobra"
+	"github.com/titantkx/titan/x/tokenfactory/types"
+)
+
+var _ = strconv.Itoa(0)
+
+func CmdMint() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "mint [amount] [mint-to-address] --from [admin] [flags]",
+		Short: "Mint a denom to an address (must have admin authority to do so)",
+		Args:  cobra.RangeArgs(1, 2),
+		RunE: func(cmd *cobra.Command, args []string) (err error) {
+			argAmount, err := sdk.ParseCoinNormalized(args[0])
+			if err != nil {
+				return err
+			}
+
+			var argReciever string
+			if len(args) == 2 {
+				argReciever = args[1]
+			}
+
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			msg := types.NewMsgMintTo(
+				clientCtx.GetFromAddress().String(),
+				argAmount,
+				argReciever,
+			)
+			if err := msg.ValidateBasic(); err != nil {
+				return err
+			}
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	flags.AddTxFlagsToCmd(cmd)
+
+	return cmd
+}