+++ x/tokenfactory/module_simulation.go
@@ -1,0 +1,162 @@
+package tokenfactory
+
+import (
+	"math/rand"
+
+	"cosmossdk.io/math"
+	"github.com/cosmos/cosmos-sdk/baseapp"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/types/module"
+	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
+	"github.com/cosmos/cosmos-sdk/x/simulation"
+	"github.com/titantkx/titan/testutil/sample"
+	tokenfactorysimulation "github.com/titantkx/titan/x/tokenfactory/simulation"
+	"github.com/titantkx/titan/x/tokenfactory/types"
+)
+
+// avoid unused import issue
+var (
+	_ = sample.AccAddress
+	_ = tokenfactorysimulation.FindAccount
+	_ = simulation.MsgEntryKind
+	_ = baseapp.Paramspace
+	_ = rand.Rand{}
+)
+
+const (
+	//nolint:gosec // this is not credentials
+	opWeightMsgCreateDenom = "op_weight_msg_create_denom"
+	// TODO: Determine the simulation weight value
+	defaultWeightMsgCreateDenom int = 100
+
+	//nolint:gosec // this is not credentials
+	opWeightMsgMint = "op_weight_msg_mint"
+	// TODO: Determine the simulation weight value
+	defaultWeightMsgMint int = 100
+
+	//nolint:gosec // this is not credentials
+	opWeightMsgBurn = "op_weight_msg_burn"
+	// TODO: Determine the simulation weight value
+	defaultWeightMsgBurn int = 100
+
+	//nolint:gosec // this is not credentials
+	opWeightMsgChangeAdmin = "op_weight_msg_change_admin"
+	// TODO: Determine the simulation weight value
+	defaultWeightMsgChangeAdmin int = 100
+
+	// this line is used by starport scaffolding # simapp/module/const
+)
+
+// GenerateGenesisState creates a randomized GenState of the module.
+func (AppModule) GenerateGenesisState(simState *module.SimulationState) {
+	accs := make([]string, len(simState.Accounts))
+	for i, acc := range simState.Accounts {
+		accs[i] = acc.Address.String()
+	}
+	tokenfactoryGenesis := types.GenesisState{
+		Params: types.DefaultParams(),
+		// this line is used by starport scaffolding # simapp/module/genesisState
+	}
+	tokenfactoryGenesis.Params.DenomCreationFee = sdk.NewCoins(sdk.NewCoin("stake", math.NewInt(10000000)))
+	simState.GenState[types.ModuleName] = simState.Cdc.MustMarshalJSON(&tokenfactoryGenesis)
+}
+
+// RegisterStoreDecoder registers a decoder.
+func (am AppModule) RegisterStoreDecoder(_ sdk.StoreDecoderRegistry) {}
+
+// ProposalContents doesn't return any content functions for governance proposals.
+func (AppModule) ProposalContents(_ module.SimulationState) []simtypes.WeightedProposalContent { //nolint:staticcheck
+	return nil
+}
+
+// WeightedOperations returns the all the gov module operations with their respective weights.
+func (am AppModule) WeightedOperations(simState module.SimulationState) []simtypes.WeightedOperation {
+	operations := make([]simtypes.WeightedOperation, 0)
+
+	var weightMsgCreateDenom int
+	simState.AppParams.GetOrGenerate(simState.Cdc, opWeightMsgCreateDenom, &weightMsgCreateDenom, nil,
+		func(_ *rand.Rand) {
+			weightMsgCreateDenom = defaultWeightMsgCreateDenom
+		},
+	)
+	operations = append(operations, simulation.NewWeightedOperation(
+		weightMsgCreateDenom,
+		tokenfactorysimulation.SimulateMsgCreateDenom(am.accountKeeper, am.bankKeeper, am.keeper),
+	))
+
+	var weightMsgMint int
+	simState.AppParams.GetOrGenerate(simState.Cdc, opWeightMsgMint, &weightMsgMint, nil,
+		func(_ *rand.Rand) {
+			weightMsgMint = defaultWeightMsgMint
+		},
+	)
+	operations = append(operations, simulation.NewWeightedOperation(
+		weightMsgMint,
+		tokenfactorysimulation.SimulateMsgMint(am.accountKeeper, am.bankKeeper, am.keeper),
+	))
+
+	var weightMsgBurn int
+	simState.AppParams.GetOrGenerate(simState.Cdc, opWeightMsgBurn, &weightMsgBurn, nil,
+		func(_ *rand.Rand) {
+			weightMsgBurn = defaultWeightMsgBurn
+		},
+	)
+	operations = append(operations, simulation.NewWeightedOperation(
+		weightMsgBurn,
+		tokenfactorysimulation.SimulateMsgBurn(am.accountKeeper, am.bankKeeper, am.keeper),
+	))
+
+	var weightMsgChangeAdmin int
+	simState.AppParams.GetOrGenerate(simState.Cdc, opWeightMsgChangeAdmin, &weightMsgChangeAdmin, nil,
+		func(_ *rand.Rand) {
+			weightMsgChangeAdmin = defaultWeightMsgChangeAdmin
+		},
+	)
+	operations = append(operations, simulation.NewWeightedOperation(
+		weightMsgChangeAdmin,
+		tokenfactorysimulation.SimulateMsgChangeAdmin(am.accountKeeper, am.bankKeeper, am.keeper),
+	))
+
+	// this line is used by starport scaffolding # simapp/module/operation
+
+	return operations
+}
+
+// ProposalMsgs returns msgs used for governance proposals for simulations.
+func (am AppModule) ProposalMsgs(_ module.SimulationState) []simtypes.WeightedProposalMsg {
+	return []simtypes.WeightedProposalMsg{
+		simulation.NewWeightedProposalMsg(
+			opWeightMsgCreateDenom,
+			defaultWeightMsgCreateDenom,
+			func(_ *rand.Rand, _ sdk.Context, _ []simtypes.Account) sdk.Msg {
+				tokenfactorysimulation.SimulateMsgCreateDenom(am.accountKeeper, am.bankKeeper, am.keeper)
+				return nil
+			},
+		),
+		simulation.NewWeightedProposalMsg(
+			opWeightMsgMint,
+			defaultWeightMsgMint,
+			func(_ *rand.Rand, _ sdk.Context, _ []simtypes.Account) sdk.Msg {
+				tokenfactorysimulation.SimulateMsgMint(am.accountKeeper, am.bankKeeper, am.keeper)
+				return nil
+			},
+		),
+		simulation.NewWeightedProposalMsg(
+			opWeightMsgBurn,
+			defaultWeightMsgBurn,
+			func(_ *rand.Rand, _ sdk.Context, _ []simtypes.Account) sdk.Msg {
+				tokenfactorysimulation.SimulateMsgBurn(am.accountKeeper, am.bankKeeper, am.keeper)
+				return nil
+			},
+		),
+		simulation.NewWeightedProposalMsg(
+			opWeightMsgChangeAdmin,
+			defaultWeightMsgChangeAdmin,
+			func(_ *rand.Rand, _ sdk.Context, _ []simtypes.Account) sdk.Msg {
+				tokenfactorysimulation.SimulateMsgChangeAdmin(am.accountKeeper, am.bankKeeper, am.keeper)
+				return nil
+			},
+		),
+		// this line is used by starport scaffolding # simapp/module/OpMsg
+	}
+}