+++ x/tokenfactory/types/denoms_test.go
@@ -1,0 +1,130 @@
+package types_test
+
+import (
+	"fmt"
+	"testing"
+
+	"github.com/stretchr/testify/require"
+
+	"github.com/titantkx/titan/testutil/sample"
+	"github.com/titantkx/titan/x/tokenfactory/types"
+)
+
+func TestDeconstructDenom(t *testing.T) {
+	expectedCreator := sample.AccAddress().String()
+
+	for _, tc := range []struct {
+		desc             string
+		denom            string
+		expectedSubdenom string
+		err              error
+	}{
+		{
+			desc:  "empty is invalid",
+			denom: "",
+			err:   types.ErrInvalidDenom,
+		},
+		{
+			desc:             "normal",
+			denom:            fmt.Sprintf("factory/%s/bitcoin", expectedCreator),
+			expectedSubdenom: "bitcoin",
+		},
+		{
+			desc:             "multiple slashes in subdenom",
+			denom:            fmt.Sprintf("factory/%s/bitcoin/1", expectedCreator),
+			expectedSubdenom: "bitcoin/1",
+		},
+		{
+			desc:             "no subdenom",
+			denom:            fmt.Sprintf("factory/%s/", expectedCreator),
+			expectedSubdenom: "",
+		},
+		{
+			desc:  "incorrect prefix",
+			denom: fmt.Sprintf("ibc/%s/bitcoin", expectedCreator),
+			err:   types.ErrInvalidDenom,
+		},
+		{
+			desc:             "subdenom of only slashes",
+			denom:            fmt.Sprintf("factory/%s/////", expectedCreator),
+			expectedSubdenom: "////",
+		},
+		{
+			desc:  "too long name",
+			denom: fmt.Sprintf("factory/%s/adsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsf", expectedCreator),
+			err:   types.ErrInvalidDenom,
+		},
+	} {
+		t.Run(tc.desc, func(t *testing.T) {
+			creator, subdenom, err := types.DeconstructDenom(tc.denom)
+			if tc.err != nil {
+				require.ErrorContains(t, err, tc.err.Error())
+			} else {
+				require.NoError(t, err)
+				require.Equal(t, expectedCreator, creator)
+				require.Equal(t, tc.expectedSubdenom, subdenom)
+			}
+		})
+	}
+}
+
+func TestGetTokenDenom(t *testing.T) {
+	for _, tc := range []struct {
+		desc     string
+		creator  string
+		subdenom string
+		valid    bool
+	}{
+		{
+			desc:     "normal",
+			creator:  sample.AccAddress().String(),
+			subdenom: "bitcoin",
+			valid:    true,
+		},
+		{
+			desc:     "multiple slashes in subdenom",
+			creator:  sample.AccAddress().String(),
+			subdenom: "bitcoin/1",
+			valid:    true,
+		},
+		{
+			desc:     "no subdenom",
+			creator:  sample.AccAddress().String(),
+			subdenom: "",
+			valid:    true,
+		},
+		{
+			desc:     "subdenom of only slashes",
+			creator:  sample.AccAddress().String(),
+			subdenom: "/////",
+			valid:    true,
+		},
+		{
+			desc:     "too long name",
+			creator:  sample.AccAddress().String(),
+			subdenom: "adsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsfadsf",
+			valid:    false,
+		},
+		{
+			desc:     "subdenom is exactly max length",
+			creator:  sample.AccAddress().String(),
+			subdenom: "bitcoinfsadfsdfeadfsafwefsefsefsdfsdafasefsf",
+			valid:    true,
+		},
+		{
+			desc:     "creator is exactly max length",
+			creator:  "titan1tk9ahr5eann843v82z50z8mx6p7lw7dm22t5kljhgjhgkhjklhkjhkjhgjhgjgjghelug",
+			subdenom: "bitcoin",
+			valid:    true,
+		},
+	} {
+		t.Run(tc.desc, func(t *testing.T) {
+			_, err := types.GetTokenDenom(tc.creator, tc.subdenom)
+			if tc.valid {
+				require.NoError(t, err)
+			} else {
+				require.Error(t, err)
+			}
+		})
+	}
+}