+++ x/tokenfactory/types/message_update_params.go
@@ -1,0 +1,37 @@
+package types
+
+import (
+	errorsmod "cosmossdk.io/errors"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+)
+
+const TypeMsgUpdateParams = "update_params"
+
+var _ sdk.Msg = &MsgUpdateParams{}
+
+// MsgUpdateParams creates a message to update module params
+func NewMsgUpdateParams(sender string, params Params) *MsgUpdateParams {
+	return &MsgUpdateParams{
+		Authority: sender,
+		Params:    params,
+	}
+}
+
+func (m MsgUpdateParams) Route() string { return RouterKey }
+
+func (m MsgUpdateParams) Type() string { return TypeMsgUpdateParams }
+
+func (m MsgUpdateParams) ValidateBasic() error {
+	_, err := sdk.AccAddressFromBech32(m.Authority)
+	if err != nil {
+		return errorsmod.Wrapf(sdkerrors.ErrInvalidAddress, "invalid authority address (%s)", err)
+	}
+
+	return m.Params.Validate()
+}
+
+func (m MsgUpdateParams) GetSigners() []sdk.AccAddress {
+	sender, _ := sdk.AccAddressFromBech32(m.Authority)
+	return []sdk.AccAddress{sender}
+}