+++ x/validatorreward/types/message_set_rate_test.go
@@ -27,33 +27,33 @@ func TestMsgSetRate_ValidateBasic(t *testing.T) {
 		}, {
 			name: "valid address",
 			msg: MsgSetRate{
-				Authority: sample.AccAddress(),
+				Authority: sample.AccAddress().String(),
 				Rate:      sdk.NewDec(0),
 			},
 		}, {
 			name: "rate not provided",
 			msg: MsgSetRate{
-				Authority: sample.AccAddress(),
+				Authority: sample.AccAddress().String(),
 			},
 			err: ErrInvalidRate,
 		}, {
 			name: "rate too low",
 			msg: MsgSetRate{
-				Authority: sample.AccAddress(),
+				Authority: sample.AccAddress().String(),
 				Rate:      sdk.NewDec(-1),
 			},
 			err: ErrInvalidRate,
 		}, {
 			name: "rate too high",
 			msg: MsgSetRate{
-				Authority: sample.AccAddress(),
+				Authority: sample.AccAddress().String(),
 				Rate:      sdk.NewDec(1000000000000000000),
 			},
 			err: ErrInvalidRate,
 		}, {
 			name: "valid rate",
 			msg: MsgSetRate{
-				Authority: sample.AccAddress(),
+				Authority: sample.AccAddress().String(),
 				Rate:      sdk.NewDecWithPrec(1, 1),
 			},
 		},